moved compare config data to common utlity & removed commented code
